과목 데이터 sql, 온보딩 권한 없음 문제 해결, api 테스트 1차

+++ src/main/java/com/kaidey/yakchatproject/domain/onboarding/dto/StudentOnboardingRes.java
@@ -8,5 +8,5 @@ import java.util.List;
 public record StudentOnboardingRes(
         Long userId,
         boolean completed,
-        Integer profileVersion
+        Long profileVersion
 ) {}
